backend/requirements.txt
# PhishGard-AI/requirements.txt

# Core
google-api-python-client  # Pour interagir avec l'API Gmail
google-auth-oauthlib  # Pour l'authentification avec les API Google
google-auth-httplib2  # Pour l'authentification avec les API Google
python-dotenv         # Pour gérer les variables d'environnement (.env)
cryptography          # Pour le chiffrement des tokens

# Auth
python-jose[cryptography] # Pour les tokens JWT
passlib[bcrypt]           # Pour le hachage (bonne pratique)

# LLM
transformers  # Pour les modèles de langage
torch

# OSINT
requests
python-whois
python-dateutil

# URL Analysis
pandas
numpy
scikit-learn
joblib
tldextract
python-whois
beautifulsoup4
selenium

# Web Framework
uvicorn 
fastapi